requirement.txt
AWSIoTPythonSDK==1.2.0
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.5.1
Django==1.11.6
django-common-helpers==0.9.1
django-cron==0.5.0
django-filter==1.0.4
django-mathfilters==0.4.0
django-nose==1.4.5
django-user-language-middleware==0.0.2
djangorestframework==3.7.0
idna==2.6
Markdown==2.6.9
nose==1.3.7
pycron==0.40
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
twill==0.9
urllib3==1.22
UserManager==0.5.3
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
